Moved User and Team to Model folder. Created User and Team Services
