+++ Manage_tasks_Biznes_Logic/Manage_tasks_Biznes_Logic.csproj
@@ -6,9 +6,4 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="Model\" />
-    <Folder Include="Service\" />
-  </ItemGroup>
-
 </Project>
