+++ Manage_tasks_Biznes_Logic/Service/IUserService.cs
@@ -1,0 +1,5 @@
+﻿namespace Manage_tasks_Biznes_Logic.Service;
+
+public interface IUserService
+{
+}
