+++ Manage_tasks_Biznes_Logic/Service/UserService.cs
@@ -1,0 +1,6 @@
+﻿namespace Manage_tasks_Biznes_Logic.Service;
+
+public class UserService : IUserService
+{
+
+}
